+++ controllers/base_controller.rb
@@ -1,5 +1,6 @@
 class BaseController
   include RequestUtils
+  include GameUtils
 
   attr_reader :path, :template_path
   attr_accessor :request, :errors, :env
@@ -38,49 +39,5 @@ class BaseController
   def render_partial(path)
     path = File.expand_path(path)
     ERB.new(File.read(path)).result(binding)
-  end
-
-  def secret_code
-    game.secret_code
-  end
-
-  def game
-    @request.session[:game]
-  end
-
-  def hints
-    attempts_info[:hints]
-  end
-
-  def hints_shown
-    game.hints_shown
-  end
-
-  def game_hash
-    game.to_h
-  end
-
-  def attempts_info
-    game_hash[:attempts_info]
-  end
-
-  def hints_total
-    attempts_info[:hints_total]
-  end
-
-  def hints_left
-    attempts_info[:hints].length
-  end
-
-  def attempts_total
-    attempts_info[:attempts_total]
-  end
-
-  def attempts_left
-    attempts_info[:attempts_left]
-  end
-
-  def difficulty
-    game_hash[:difficulty]
   end
 end
