+++ initializer.rb
@@ -3,9 +3,13 @@ require 'delegate'
 require 'bundler/setup'
 require 'codebreaker'
 require 'time'
+require 'i18n'
+
+I18n.load_path << Dir[File.expand_path("config/locales") + "/*.yml"]
 
 require_relative './constants'
 require_relative './utils/request_utils'
+require_relative './utils/game_utils'
 require_relative './controllers/base_controller'
 require_relative './controllers/game_controller'
 require_relative './controllers/win_game_controller'
